File rearrangement: 1) Add new package "user search engine" 2) Move abstract DS IO into ABR
